booking-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
pydantic[email]==2.9.0
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.3
